clean-up for front-end structure and code of insights page; gantt chart modularization, removal of customized header behavior as it's all one page again, data set chart filter improved to not re-order datasets (which made things more confusing to look at), etc.
