+++ packages/frontend/app/(dashboard)/course/[cid]/insights/components/charts/LineChartComponent.tsx
@@ -1,24 +1,17 @@
 import {
-  ChartContainer,
-  ChartLegend,
-  ChartLegendContent,
-  ChartTooltip,
-  ChartTooltipContent,
-} from '@/app/components/ui/chart'
-import { Area, AreaChart, CartesianGrid, XAxis, YAxis } from 'recharts'
-import React, { useMemo } from 'react'
-import {
   AxisChartClasses,
   ChartComponentProps,
   LinearChartProps,
-} from '@/app/(dashboard)/course/[cid]/(insights)/insights/utils/types'
-import { generateAxisRange } from '@/app/(dashboard)/course/[cid]/(insights)/insights/utils/functions'
+} from '@/app/(dashboard)/course/[cid]/insights/utils/types'
+import React, { useMemo } from 'react'
+import { ChartContainer } from '@/app/components/ui/chart'
+import { CartesianGrid, LineChart, Line } from 'recharts'
 import {
   getAxisComponents,
   getLegendAndTooltipComponents,
-} from '@/app/(dashboard)/course/[cid]/(insights)/insights/components/charts/ChartFunctions'
+} from '@/app/(dashboard)/course/[cid]/insights/components/charts/ChartFunctions'
 
-const AreaChartComponent: React.FC<ChartComponentProps> = ({ props }) => {
+const LineChartComponent: React.FC<ChartComponentProps> = ({ props }) => {
   const {
     chartConfig,
     chartData,
@@ -41,7 +34,7 @@ const AreaChartComponent: React.FC<ChartComponentProps> = ({ props }) => {
     tickLine,
     tickMargin,
     axisLine,
-    axisRatio,
+    aspectRatio,
     tickFormatter,
   } = props as LinearChartProps
 
@@ -53,7 +46,7 @@ const AreaChartComponent: React.FC<ChartComponentProps> = ({ props }) => {
   tickLine ??= true
   tickMargin ??= 8
   axisLine ??= true
-  axisRatio ??= 2
+  aspectRatio ??= 2
   tickFormatter ??= (value) => (value as string).substring(0, 3)
 
   const className = useMemo(() => {
@@ -66,9 +59,9 @@ const AreaChartComponent: React.FC<ChartComponentProps> = ({ props }) => {
     <ChartContainer
       config={chartConfig}
       className={className}
-      style={{ aspectRatio: axisRatio }}
+      style={{ aspectRatio: aspectRatio }}
     >
-      <AreaChart data={chartData} accessibilityLayer>
+      <LineChart data={chartData}>
         <CartesianGrid vertical={verticalAxis} />
         {getAxisComponents(
           chartData,
@@ -88,42 +81,20 @@ const AreaChartComponent: React.FC<ChartComponentProps> = ({ props }) => {
           valueFormatter,
           legendFormatter,
         )}
-        <defs>
-          {valueKeys &&
-            valueFills &&
-            valueKeys.map((key) => (
-              <>
-                <linearGradient id={'fill' + key} x1="0" y1="0" x2="0" y2="1">
-                  <stop
-                    offset={'5%'}
-                    stopColor={valueFills[key]}
-                    stopOpacity={0.8}
-                  />
-                  <stop
-                    offset={'95%'}
-                    stopColor={valueFills[key]}
-                    stopOpacity={0.0}
-                  />
-                </linearGradient>
-              </>
-            ))}
-        </defs>
         {valueKeys &&
           valueFills &&
           valueKeys.map((key, index) => (
-            <Area
+            <Line
               key={index}
               stroke={valueFills[key]}
-              strokeWidth={2}
-              fill={`url(#fill${key})`}
               dataKey={key}
               dot={showPoints}
               type={curveType ?? 'monotone'}
             />
           ))}
-      </AreaChart>
+      </LineChart>
     </ChartContainer>
   )
 }
 
-export default AreaChartComponent
+export default LineChartComponent
