+++ packages/frontend/app/(dashboard)/course/[cid]/insights/components/filters/FilterWrapper.tsx
