+++ packages/frontend/app/(dashboard)/course/[cid]/insights/context/InsightsContext.tsx
